Turn RovrRenderer into the first post-render plugin
Instead of it being the last pre-render plugin, it's now the first
post-render plugin.

Both pre and post don't really make sense for this plugin since it
literally is the renderer. But making it the first post function allows
it to be added immediately in the constructor rather than later on.

Also, I moved where the deletion of the destination directory happens
to make it more clear.