Allow code blocks with no syntax highlighting
A previous decision of mine disallowed code blocks without a specified
language to prevent the possibility of another bug happening.

This was a bad decision because code blocks without languages should
definitely be allowed as per the markdown specification.

The bug I was trying to prevent occurs when `marked` parses markdown
that contains html code with line breaks wrapped within a div. If the
content within that div is tabbed, `marked` interprets it as a code
block which is understandable. This should be considered a minor bug.
It even happens within Atom's markdown preview plugin.